pyproject.toml
[project]
name = "simple-django-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Trung Nguyen",email = "ng.qtrung95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-tailwind[reload] (>=4.0.1,<5.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "dj-database-url (>=3.0.0,<4.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "django-tinymce (>=4.1.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-taggit (>=6.1.0,<7.0.0)",
    "django-storages[s3] (>=1.14.6,<2.0.0)",
    "defusedxml (>=0.7.1,<0.8.0)",
    "django-taggit-labels (>=0.7.0,<0.8.0)",
    "django-active-link (>=0.2.2,<0.3.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "weasyprint (>=66.0,<67.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "redis[hiredis] (>=6.4.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "discord-py (>=2.6.0,<3.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
create-django = "my_module:main"
